+++ packages/building-rollup/demo/ts-babel/demo-app.ts
@@ -1,7 +1,7 @@
-import { LitElement, html } from 'lit-element';
+import { LitElement, html, TemplateResult } from 'lit-element';
 
 class DemoApp extends LitElement {
-  render() {
+  render(): TemplateResult {
     return html`
       <h1>Hello typescript!</h1>
     `;