+++ packages/building-webpack/modern-and-legacy-config.js
@@ -1,37 +1,39 @@
 const { findSupportedBrowsers } = require('@open-wc/building-utils');
 const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+const WebpackIndexHTMLPlugin = require('@open-wc/webpack-index-html-plugin');
 const CleanWebpackPlugin = require('clean-webpack-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
-const ModernWebWebpackPlugin = require('./modern-web-webpack-plugin');
+const merge = require('webpack-merge');
 
 const development = !process.argv.find(arg => arg.includes('production'));
 const legacy = process.argv.find(arg => arg.includes('legacy'));
 
-const prefix = '[@open-wc/building-webpack/modern-and-legacy-config]:';
-const { queryAll, predicates, getAttribute } = require('./dom5-fork/index.js');
-const { readHTML } = require('./src/utils.js');
-
-const modernWebWebpackPlugin = new ModernWebWebpackPlugin({ development });
-
 const defaultOptions = {
-  indexHTML: './index.html',
-  entry: './index.js',
-  htmlEntryPoint: false,
+  input: './index.html',
 };
 
 /* eslint-disable-next-line no-shadow */
 function createConfig(options, legacy) {
+  if (options.entry) {
+    /* eslint-disable-next-line no-console */
+    console.warn(
+      '[@open-wc/building-webpack]: options.entry is deprecated. Use options.input instead.',
+    );
+  }
+
+  const inputPrefix = legacy ? 'legacy/' : '';
+  const outputFilename = `${inputPrefix}[name].${development ? '' : '[contenthash].'}js`;
+
   return {
-    entry: Array.isArray(options.entry) ? options.entry : [options.entry],
+    entry: options.input || options.entry,
 
     output: {
-      filename: `${legacy ? 'legacy/' : ''}[name].[chunkhash].js`,
-      chunkFilename: `${legacy ? 'legacy/' : ''}[name].[chunkhash].js`,
+      filename: outputFilename,
+      chunkFilename: outputFilename,
       path: path.resolve(process.cwd(), `dist`),
     },
 
-    devtool: development ? 'inline-source-map' : 'source-map',
+    devtool: development ? 'cheap-module-source-map' : 'source-map',
 
     resolve: {
       mainFields: [
@@ -67,6 +69,7 @@ function createConfig(options, legacy) {
                       collapseWhitespace: true,
                       removeComments: true,
                       caseSensitive: true,
+                      minifyCSS: true,
                     },
                   },
                 ],
@@ -88,10 +91,6 @@ function createConfig(options, legacy) {
               ],
             },
           },
-        },
-        options.htmlEntryPoint && {
-          test: options.input,
-          loader: require.resolve('./src/clean-up-html-loader.js'),
         },
       ].filter(_ => !!_),
     },
@@ -115,12 +114,21 @@ function createConfig(options, legacy) {
       // @ts-ignore
       !development && new CleanWebpackPlugin(),
 
-      new HtmlWebpackPlugin({
-        template: options.indexHTML,
-        inject: false,
-      }),
-
-      modernWebWebpackPlugin,
+      new WebpackIndexHTMLPlugin(
+        merge(
+          {
+            multiBuild: !development,
+            legacy,
+            polyfills: {
+              coreJs: true,
+              regeneratorRuntime: true,
+              webcomponents: true,
+              fetch: true,
+            },
+          },
+          options.webpackIndexHTMLPlugin,
+        ),
+      ),
     ].filter(_ => !!_),
 
     devServer: {
@@ -140,28 +148,6 @@ module.exports = userOptions => {
     ...defaultOptions,
     ...userOptions,
   };
-
-  if (typeof options.input === 'string' && options.input.endsWith('index.html')) {
-    options.indexHTML = options.input;
-    options.htmlEntryPoint = true;
-    const indexHTML = readHTML(options.input);
-    const scripts = queryAll(indexHTML, predicates.hasTagName('script'));
-    const moduleScripts = scripts.filter(script => getAttribute(script, 'type') === 'module');
-
-    if (moduleScripts.length === 0) {
-      throw new Error(
-        `${prefix} Could not find any module script in configured input: ${options.input}`,
-      );
-    }
-
-    if (moduleScripts.some(script => !getAttribute(script, 'src'))) {
-      throw new Error(`${prefix} Module scripts without a 'src' attribute are not supported.`);
-    }
-    const indexDir = path.dirname(options.input);
-
-    const modules = moduleScripts.map(script => getAttribute(script, 'src'));
-    options.entry = modules.map(p => path.join(indexDir, p));
-  }
 
   if (development) {
     return createConfig(options, legacy);