+++ packages/testing-helpers/test/fixture.test.js
@@ -34,11 +34,13 @@ describe('fixtureSync & fixture', () => {
     const elementSync = fixtureSync(html`
       <div .propNumber=${10} .propFunction=${myFunction}></div>
     `);
+    // @ts-ignore
     testElement(elementSync);
 
     const elementAsync = await fixture(html`
       <div .propNumber=${10} .propFunction=${myFunction}></div>
     `);
+    // @ts-ignore
     testElement(elementAsync);
   });