+++ packages/webpack-index-html-plugin/src/utils.js
@@ -1,0 +1,11 @@
+const PLUGIN_NAME = 'WebpackIndexHTMLPlugin';
+
+/** @param {string} msg */
+function createError(msg) {
+  return new Error(`[${PLUGIN_NAME}]: ${msg}`);
+}
+
+module.exports = {
+  PLUGIN_NAME,
+  createError,
+};