+++ packages/building-rollup/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.3](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.5.2...@open-wc/building-rollup@0.5.3) (2019-06-13)
+
+
+### Bug Fixes
+
+* **building-rollup:** clean falsy plugins in legacy watch ([#490](https://github.com/open-wc/open-wc/issues/490)) ([58d8e63](https://github.com/open-wc/open-wc/commit/58d8e63))
+* **building-rollup:** make ts builds work for modern ([#491](https://github.com/open-wc/open-wc/issues/491)) ([e940904](https://github.com/open-wc/open-wc/commit/e940904))
+
+
+
+
+
+## [0.5.2](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.5.1...@open-wc/building-rollup@0.5.2) (2019-06-13)
+
+
+### Bug Fixes
+
+* **building-rollup:** remove legacy scripts from modern watch mode ([#489](https://github.com/open-wc/open-wc/issues/489)) ([8083431](https://github.com/open-wc/open-wc/commit/8083431))
+
+
+
+
+
+## [0.5.1](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.5.0...@open-wc/building-rollup@0.5.1) (2019-06-12)
+
+
+### Bug Fixes
+
+* **building-rollup:** enable and show howto setup typescript ([36d9325](https://github.com/open-wc/open-wc/commit/36d9325))
+
+
+
+
+
+# [0.5.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.4.7...@open-wc/building-rollup@0.5.0) (2019-06-02)
+
+
+### Features
+
+* **building-rollup:** include all static modules in `entryModules` ([#473](https://github.com/open-wc/open-wc/issues/473)) ([2dec7a2](https://github.com/open-wc/open-wc/commit/2dec7a2))
+
+
+
+
+
+## [0.4.7](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.4.6...@open-wc/building-rollup@0.4.7) (2019-05-25)
+
+
+### Bug Fixes
+
+* **building-rollup:** minify CSS properly ([bee9086](https://github.com/open-wc/open-wc/commit/bee9086))
+
+
+
+
+
 ## [0.4.6](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.4.5...@open-wc/building-rollup@0.4.6) (2019-05-20)
 
 **Note:** Version bump only for package @open-wc/building-rollup