+++ packages/building-webpack/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.5](https://github.com/open-wc/open-wc/compare/@open-wc/building-webpack@1.4.4...@open-wc/building-webpack@1.4.5) (2019-05-30)
+
+
+### Bug Fixes
+
+* **building-webpack:** html race condition ([10c22ab](https://github.com/open-wc/open-wc/commit/10c22ab))
+
+
+
+
+
+## [1.4.4](https://github.com/open-wc/open-wc/compare/@open-wc/building-webpack@1.4.3...@open-wc/building-webpack@1.4.4) (2019-05-25)
+
+
+### Bug Fixes
+
+* **building-webpack:** minify CSS properly ([81fe2c2](https://github.com/open-wc/open-wc/commit/81fe2c2))
+
+
+
+
+
 ## [1.4.3](https://github.com/open-wc/open-wc/compare/@open-wc/building-webpack@1.4.2...@open-wc/building-webpack@1.4.3) (2019-05-19)
 
 **Note:** Version bump only for package @open-wc/building-webpack