+++ packages/chai-dom-equals/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.6](https://github.com/open-wc/open-wc/compare/@open-wc/chai-dom-equals@0.12.5...@open-wc/chai-dom-equals@0.12.6) (2019-06-08)
+
+**Note:** Version bump only for package @open-wc/chai-dom-equals
+
+
+
+
+
+## [0.12.5](https://github.com/open-wc/open-wc/compare/@open-wc/chai-dom-equals@0.12.4...@open-wc/chai-dom-equals@0.12.5) (2019-05-25)
+
+**Note:** Version bump only for package @open-wc/chai-dom-equals
+
+
+
+
+
 ## [0.12.4](https://github.com/open-wc/open-wc/compare/@open-wc/chai-dom-equals@0.12.3...@open-wc/chai-dom-equals@0.12.4) (2019-05-19)
 
 **Note:** Version bump only for package @open-wc/chai-dom-equals