+++ packages/create/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.3](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.12.2...@open-wc/create@0.12.3) (2019-06-05)
+
+
+### Bug Fixes
+
+* **create:** use correct open path ([02cab4b](https://github.com/open-wc/open-wc/commit/02cab4b))
+
+
+
+
+
 ## [0.12.2](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.12.1...@open-wc/create@0.12.2) (2019-05-20)