+++ packages/webpack-import-meta-loader/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-import-meta-loader@0.2.3...@open-wc/webpack-import-meta-loader@0.2.4) (2019-06-05)
+
+**Note:** Version bump only for package @open-wc/webpack-import-meta-loader
+
+
+
+
+
+## [0.2.3](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-import-meta-loader@0.2.2...@open-wc/webpack-import-meta-loader@0.2.3) (2019-06-04)
+
+**Note:** Version bump only for package @open-wc/webpack-import-meta-loader
+
+
+
+
+
+## [0.2.2](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-import-meta-loader@0.2.1...@open-wc/webpack-import-meta-loader@0.2.2) (2019-06-03)
+
+
+### Bug Fixes
+
+* **webpack-import-meta-loader:** rewrite all backslashes ([#477](https://github.com/open-wc/open-wc/issues/477)) ([ade8a65](https://github.com/open-wc/open-wc/commit/ade8a65))
+
+
+
+
+
 ## [0.2.1](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-import-meta-loader@0.2.0...@open-wc/webpack-import-meta-loader@0.2.1) (2019-05-03)